+++ src/tribler-core/tribler_core/components/bandwidth_accounting/bandwidth_accounting_component.py
