+++ src/tribler-core/tribler_core/components/tunnels.py
@@ -2,8 +2,8 @@ from ipv8.dht.provider import DHTCommunityProvider
 from ipv8.messaging.anonymization.community import TunnelSettings
 from ipv8.peerdiscovery.discovery import RandomWalk
 from ipv8_service import IPv8
+from tribler_core.components.bandwidth_accounting.bandwidth_accounting_component import BandwidthAccountingComponent
 from tribler_core.components.base import Component
-from tribler_core.components.bandwidth_accounting.bandwidth_accounting import BandwidthAccountingComponent
 from tribler_core.components.ipv8 import Ipv8Component
 from tribler_core.components.libtorrent import LibtorrentComponent
 from tribler_core.components.reporter import ReporterComponent
