+++ src/tribler-core/tribler_core/modules/tests/bandwidth_accounting/test_community.py
@@ -3,12 +3,11 @@ from asyncio import sleep
 from ipv8.test.base import TestBase
 from ipv8.test.mocking.ipv8 import MockIPv8
 
-from tribler_core.modules.bandwidth_accounting import EMPTY_SIGNATURE
 from tribler_core.components.bandwidth_accounting.community.cache import BandwidthTransactionSignCache
 from tribler_core.components.bandwidth_accounting.community.community import BandwidthAccountingCommunity
 from tribler_core.components.bandwidth_accounting.db.database import BandwidthDatabase
 from tribler_core.components.bandwidth_accounting.settings import BandwidthAccountingSettings
-from tribler_core.components.bandwidth_accounting.db.transaction import BandwidthTransactionData
+from tribler_core.components.bandwidth_accounting.db.transaction import BandwidthTransactionData, EMPTY_SIGNATURE
 from tribler_core.utilities.utilities import MEMORY_DB
