+++ src/tribler-core/tribler_core/restapi/tests/test_trustview_endpoint.py
@@ -9,9 +9,8 @@ from ipv8.keyvault.crypto import default_eccrypto
 import pytest
 
 from tribler_core.exceptions import TrustGraphException
-from tribler_core.modules.bandwidth_accounting import EMPTY_SIGNATURE
 from tribler_core.components.bandwidth_accounting.db.database import BandwidthDatabase
-from tribler_core.components.bandwidth_accounting.db.transaction import BandwidthTransactionData
+from tribler_core.components.bandwidth_accounting.db.transaction import BandwidthTransactionData, EMPTY_SIGNATURE
 from tribler_core.modules.trust_calculation.trust_graph import TrustGraph
 from tribler_core.restapi.base_api_test import do_request
 from tribler_core.restapi.rest_manager import error_middleware
