+++ src/tribler-core/tribler_core/start_core.py
@@ -13,7 +13,7 @@ from tribler_common.version_manager import VersionHistory
 
 import tribler_core
 from tribler_core.check_os import check_and_enable_code_tracing, set_process_priority
-from tribler_core.components.bandwidth_accounting.bandwidth_accounting import BandwidthAccountingComponent
+from tribler_core.components.bandwidth_accounting.bandwidth_accounting_component import BandwidthAccountingComponent
 from tribler_core.components.base import Component, Session
 from tribler_core.components.gigachannel import GigaChannelComponent
 from tribler_core.components.gigachannel_manager import GigachannelManagerComponent
